refactor: Restructure web crawler utilities and models
- Removed HTMLParser and TableParser from the web crawler utils module, consolidating functionality into a new models module.
- Introduced CrawledDataStorage to aggregate receipt HTML with parsing capabilities for metadata and product data.
- Updated Crawler class to return CrawledDataStorage instead of raw HTML, enhancing data handling.
- Adjusted __all__ exports to reflect the new structure and ensure proper module organization.
