+++ backend/src/expenses_counter/utils/web_crawler/__init__.py
@@ -1,7 +1,6 @@
 """Web crawler utils module."""
 
-from .crawler import Crawler
-from .html_parser import HTMLParser
-from .table_parser import TableParser
+__all__ = ("CrawledDataStorage", "Crawler")
 
-__all__ = ("Crawler", "HTMLParser", "TableParser")
+from .crawler import Crawler
+from .models import CrawledDataStorage
