+++ backend/src/expenses_counter/utils/web_crawler/models/__init__.py
@@ -1,0 +1,7 @@
+"""Models module."""
+
+__all__ = ("CrawledDataStorage", "HTMLParser", "TableParser")
+
+from .crawled_data_storage import CrawledDataStorage
+from .html_parser import HTMLParser
+from .table_parser import TableParser
